adding back if

+++ setup.bashrc
@@ -7,13 +7,13 @@ chmod 700 win.bashrc
 curl -s -L -O https://raw.githubusercontent.com/yeltnar/gist/master/mac.bashrc
 chmod 700 mac.bashrc
 
-#if ! [ -f ~/.bashrc ]; then
+if ! [ -f ~/.bashrc ]; then
 	touch ~/.bashrc
-#fi
+fi
 
-#if [ -f ~/.bash_profile ]; then
+if [ -f ~/.bash_profile ]; then
 	echo -e "$(cat ~/.bash_profile)\nif [ -f ~/.bashrc ]; then . ~/.bashrc; fi" > ~/.bash_profile
-#fi
+fi
 
 if ! [[ $(cat ~/.bashrc) =~ ^.*\.bashrc.*$ ]] ; 
 then
